Configure devise for users
This PR Sets up devise for users. First, ran rails generate devise:install and then
added the mailcatcher gem. Added mailcatcher configurations to the
/config/environments/developmnent.rb to set up the mailer port and host.
The mailer port and host are assigned by using ENV files located in the
new .env file. An .env.example was also created so that anyone can
figure out what variables need to be set. Then this change used generated
controllers and views for devise user.
Next this creates the /users/views for edit and new. It also updates
the migrations for user devise model. Next Generate views for users/session,
and update routes to use /users/sessions. This is done because we are
overriding the default devise controllers.
The views for users session include sign in, gives the route to
sign out and delete the session, and also utilizes the navigation bar
to handle if the drop down says sign in or sign out based on the current
users state.