Push email out to Sidekiq
This way, we aren't sending the email during the request/response cycle,
delaying (and possibly breaking) the HTTP response. The email will send
whenever we have capacity for it and if it breaks it doesn't interrupt
the HTTP response. It will be retried through Sidekiq's retry mechanism.

This commit also moves the check for `ok_to_email` inside a check for a
saved record. We've already run `volunteer.save` so we don't need to run
it again.