Create user management (#114)
* Generate user_management_controller, create views

* Create user table and config devise_invitable

This PR configures the gem `devise_invitable`. With this change, a new
model devise method of :invitable is added, and a new controller named
invitations_controller exists in /users. This controller uses two after
actions to redirect users to the appropriate route. The admin sidebar
was given an additional tab of `users` which will redirect to the
user_management#show page. This page will list all emails, and a boolean
which indicates whether the user has accapted and invite and has been
sent and invitation. The #show page includes an `invite` button that
will take the user to /users/invitation/new where an email can be
provided, and an email will be sent. This email will contain a link,
with token, that when clicked will redirect users to create a
password, afterwhich they will be redirected to the volunteer creation
form, to set up data. Within this form, the user select option has been
removed and replaced by setting the current user to the volunteer, done
within the volunteer model.

* ran bundle install

* drop duplicate table description

* add has_one :volunteer to user

* drop user_id requirement from volunteer

* put user_id back in

* drop extra mail block