Update BoxRequests#index markup with assocations
BoxRequest has a User association via :reviewed_by and Box has
associations to User via :designed_by, :assembled_by, and :shipped_by.
The index table shows the names of each.

I didn't find a way to confirm whether a followup was sent (it may not
be implemented yet), so it is currently hardcoded to `false`.