add a docker-compose config for services
This adds a docker-compose config that a developer can opt to use for
running the services on which the application depends, e.g. the
database, the job queue store, and a dev-only fake mailserver.

The mailcatcher image[1] from tophfr seems reasonable. It's small and
does what we want when it comes to running a Mailcatcher process to ...
catch mail being sent in development.

Also rearranged and updated the development environment requirements in
the README. Relaxed the versions of components so that they are "this
version or greater."

Provided some links to the Heroku CLI and Docker Desktop.

[1] https://hub.docker.com/r/tophfr/mailcatcher

Signed-off-by: Robb Kidd <robb@thekidds.org>