Add warden test helpers, add rspec extended output
This commit includes the Warden::test::helpers to the rspec
configuration within spec/rails_helpers.rb. This will allow us the
use of Warden's "login_as" method. This method, and configuration
should be used when testing /spec/request tests. The implementation
of this method can be seen within the /spec/requests/attendances_spec.rb
file, and documenation for this change exists under the Testing header
of the README.md file.

This commit also fixes one of the request specs to showcase the use of
the warden test helper's "login_as" method.

This commit includes the line "--format documentation" within the .rspec
file, which will produce the rspec test's "describe" and "it" blocks in
the output of the rspec tests in a human readable format.