set PG service to create a user for dev
This change should have the effect of bringing up a PostgreSQL service
that has a database user already created within it whose username
matches the developer's username on their computer.

The postgres docker image will automatically create a user within the
database service. The name for that user is set by the environment
variable POSTGRES_USER, with a default value of "postgres". So here,
we set POSTGRES_USER to the value of whatever the host OS's USER
environment variable is. This is usually the username of the developer
running docker-compose to start containers. For most UNIX-like operating
systems, this should mean that the default database connection config
should Just Work™.

For the dirty details about how this is happening under the hood, see
that the postgres image's Dockerfile[1] has an entrypoint script it runs
whenever the image is started for the first time. It is that entrypoint
script that initializes the database with a username given in
POSTGRES_USER.[2]

[1] https://github.com/docker-library/postgres/blob/9d8e2448436b2af1ea715822c2d209d493760007/11/Dockerfile#L173
[2] https://github.com/docker-library/postgres/blob/9d8e2448436b2af1ea715822c2d209d493760007/11/docker-entrypoint.sh#L75

Signed-off-by: Robb Kidd <robb@thekidds.org>