forward services' ports to loopback/localhost
The 'ports:' entry is where we can declare forwarding ports from the
running containers to elsewhere. The syntax is pretty terse.[1]

Generally, the pattern is "HOST:CONTAINER" so that what is on the right
of the colon (:) is the port on the container and what is on the left is
the port on the host running the container. But these have *two* colons!
Yup. It's still roughly host-first-then-container.

Let's take a look at one of the Mailcatcher port forwards:

    "127.0.0.1:1080:80"

The mailcatcher image "exposes" ports 80 and 25[2] because the
mailcatcher service does webby and maily things by listening on those
ports by default. For a developer to be able to communicate with
that service on those ports, we need to forward them outside of the
wee little container network docker-compose will create and link the
ports to ports on the development host. We shouldn't forward them to
80 or 25 because those low ports (port numbers less than 1024) are
reserved for privileged processes and will make your life difficult as a
developer if you try to use them. So we'll use 1080 and 1025.

    "1080:80"

This would work to forward the container's port 80 to the host's port
1080. Now you can point your usual web browser to localhost:1080 and see
Mailcatcher's web interface. But it would bind the container's port 80
to *every* network interface on the host computer. Meaning mailcatcher
would, for example, be serving itself up on your wifi interface to
everybody else in the coffee shop where you're co-working.

    "127.0.0.1:1080:80"

This says bind the container's port 80 to port 1080 on the host's
*loopback* interface (127.0.0.1) which is also 99.999% of the time what
the name "localhost" is referring to. Only other processes on your
computer will be able to communicate with loopback. Now you're not
exposing services you're running for development to the entire coffee
shop.

[1] https://docs.docker.com/compose/compose-file/#ports
[2] https://github.com/tophfr/docker-mailcatcher/blob/118cdb9e3ba53464f595c25d76413dc4e139db4e/build/Dockerfile#L24

Signed-off-by: Robb Kidd <robb@thekidds.org>