make mailcatcher internal and external ports match
The output from the mailcatcher processes are confusing when the ports
the container uses do not match the ports being forwarded. For example:

```
$ docker-compose up mailcatcher
mailcatcher_1  | Starting MailCatcher
mailcatcher_1  | ==> smtp://0.0.0.0:25
mailcatcher_1  | ==> http://0.0.0.0:80/
```

Mailcatcher has told the developer to connect to ports 25 and 80. But
we've forwarded those ports to 1025 and 1080. So, this feedback to the
developer is misleading. The ports that mailcatcher uses are specified
by the command given for the image by the Dockerfile.[1]

So here, we're overriding that command with one that looks a lot like
it, but with smtp-port and http-port set to the ports we are fowarding.
Then, the port forwards are updated now that the container internal
ports match the host external ports.

Now the mailcatcher output looks like:

```
$ docker-compose up mailcatcher
mailcatcher_1  | Starting MailCatcher
mailcatcher_1  | ==> smtp://0.0.0.0:1025
mailcatcher_1  | ==> http://0.0.0.0:1080/
```

Which provides links that actually work in the development environment.

[1] https://github.com/tophfr/docker-mailcatcher/blob/118cdb9e3ba53464f595c25d76413dc4e139db4e/build/Dockerfile#L26

Signed-off-by: Robb Kidd <robb@thekidds.org>