Create attendance policy, try to use it in tests
Currently the attendance policy exists, but is not being used within
the requests/attendance_spec.rb context of "WITHOUT permission." which
is returning the success status of 200, where we are expecting a 401.
Commented out configuration for setting up capybara with selenium web
driver has been added to the rails_helper and spec_helper files.