Update the read me to include troublshooting tips
The following update to the readme includes tips on how to kill PIDs
that prevent you from starting the application. I believe this is
important to include because I have run into this issue multiple time,
as did many others while working at Ruby for Good. It may be beneficial
to include other troubleshooting advice that often appears to help aid
those just starting out on the project.