Sets up passing rubocop using autocorrect and todo_config
Background Explanation:
- Rubocop currently has over 4000 warnings when run against the app.
In order to run a linter on future PRs, we need rubocop to pass when run
against the repository. This commit aims to resolve all of the "simpler"
rubocop warnings, and creates a rubocop_todo.yml file to clean up more
challenging warnings in the future.

This commit:
- Runs prettier and rubocop auto-correct against the repository to reduce
rubocop warnings from 4000 to 600
- Creates a ruboco_todo file to clean up remaining warnings at a later date

Documentation on tools used:
- https://rubocop.readthedocs.io/en/latest/auto_correct/
- https://rubocop.readthedocs.io/en/latest/configuration/
- https://github.com/prettier/plugin-ruby